lab_guiado/Docker_K8s/aula8/03-lab_api_redis_worker/worker/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.8